Add grave/acute as tap-layers sending ALT codes.
The two new layers are accessed via holding the respective keys in
the Workman base layout (tapping sends the usual keystroke). These
layers are opaque to everything except aeioucl, corresponding to the
catalan accented vowels, plus ce trencada and ele geminada. Might add
diacritic in its own layer in an upcoming revision.

The handling code detects both shift-keys and caps lock, and adjusts the
ALT code sent accordingly. If I ever have the need to work on Macs (and
access to one to test it on) I'll write Mac-compatible handling.